refactor(update): break AppUpdate require cycle by extracting strategy + useDownloadPackage
The previous split left useAppUpdate.tsx and AppUpdateForeground.tsx in
a require cycle that metro warned about on every cold-launch:

  Require cycle: useAppUpdate.tsx
              -> AppUpdateForeground.tsx
              -> useAppUpdate.tsx

The cycle existed in both directions:
  useAppUpdate.tsx → AppUpdateForeground.tsx (useAppUpdateForegroundEffects)
  AppUpdateForeground.tsx → useAppUpdate.tsx (useDownloadPackage,
                                              isAutoUpdateStrategy,
                                              isForceUpdateStrategy)

Pull the leaves the wrong-direction edge depended on into their own
files so AppUpdateForeground no longer touches useAppUpdate:

- updateStrategy.ts (new): pure predicates over EUpdateStrategy /
  EAppUpdateStatus — isAutoUpdateStrategy, isForceUpdateStrategy,
  isShowAppUpdateUIWhenUpdating, isShowToastError. Leaf module, only
  depends on @onekeyhq/shared.
- useDownloadPackage.tsx (new): the full download/verify/install
  orchestration hook (~400 lines), extracted verbatim. Imports from
  updateMutex, updateRetry, updateAnalytics, updateErrorTaxonomy, and
  updateStrategy — none of which reach back into useAppUpdate.

useAppUpdate.tsx now only contains useAppUpdateInfo + useAppChangeLog
and re-exports the moved symbols so existing consumers and the test
suite import paths keep working unchanged.

AppUpdateForeground.tsx imports useDownloadPackage from
./useDownloadPackage and the strategy predicates from ./updateStrategy
directly. Dependency direction is now strictly:

  index.ts → useAppUpdate.tsx → AppUpdateForeground.tsx
                              → useDownloadPackage.tsx
                              → updateStrategy.ts

All 165 tests across 4 suites still pass — the public API surface (via
index.ts) is byte-identical, so MoreActionButton / Bootstrap / views/
AppUpdate / etc. consumers don't need to change.
